[S3ClientUploader] Extract as trait, renaming implementations
