+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientMultiPartTransferManager.scala
@@ -1,0 +1,5 @@
+package net.kemitix.s3thorp.awssdk
+
+class S3ClientMultiPartTransferManager {
+
+}
