+++ src/main/scala/net/kemitix/s3thorp/awssdk/ThorpS3Client.scala
@@ -14,7 +14,7 @@ private class ThorpS3Client(s3Client: S3CatsIOClient)
   lazy val amazonS3Client = AmazonS3ClientBuilder.defaultClient()
   lazy val objectLister = new S3ClientObjectLister(s3Client)
   lazy val copier = new S3ClientCopier(s3Client)
-  lazy val uploader = new S3ClientUploader(s3Client)
+  lazy val uploader = new S3ClientPutObjectUploader(s3Client)
   lazy val multiPartUploader = new S3ClientMultiPartUploader(amazonS3Client)
   lazy val deleter = new S3ClientDeleter(s3Client)
